[project @ fixes IntegerRing().quotient(p) bugs] In a previous patch I overloaded IntegerRing().quotient(p) to be defined as IntegerModRing(p). This was done for efficiency reasons.
This broke some tests because, as it turns out, some code expects
R.quotient(I) to accept an ideal as input, and also expect the output
to be an instance of QuotientRing_generic.

The present patch fixes this in the following way:

1) adds code in IntegerRing().quotient(p) to test for p being an
ideal; if so p.gen() replaces p

2) makes IntegerModRing_generic to derive from QuotientRing_generic
(rather than CommutativeRing).

3) renames QuotientRing_generic.__repr__ to _repr_ so that the
overloading works, since IntegerModRing_generic only defines _repr_,
as expected (?) for a subclass of SageObject.

4) changes examples to account for the change in representation of the
quotient rings of IntegerRing().