Fix really stupid bug introduced at the last minute
When reducing frob(x^i dx/y) we know it has only even powers of y, so we can skip the odd ones, but it wasn't remembering that it had skipped them when constructing f... Argh.

Consistancy checks from Kiran Kedlaya now all working, which is a very good sign.