Trac #9310: Trac #9310: random doctest failures in matrix_group.py
The module_composition_factors(),  as_permutation_group(), and
random_element() methods depend on the architecture and evaluation
order of the doctests, despite us resetting the random number
generators. This patch sorts the output of module_composition_factors
and changes the doctests for the other two methods to be insensitive to
the random choices.