Trac #10039: Trac #10039: Change minimize/maximize to only return relevant information
I changed return value of the maximize/minimize methods to only return
{'bounded':False} if the linear program is unbounded, since all other
dictionary entries are not well-defined in that case. If the linear program
is bounded then the returned dictionary is the same as before and
contains information about the sup/inf value, whether it is a maximum,
and where it is attained.