Trac #9130: trac 9130: fix segfault in sage.symbolic.pynac.py_float
When the second argument of py_float() is declared as object, Cython assumes
that this is initialized and it can do reference counting, etc. Calls to
py_float() from pynac cannot provide a valid python object for the second
argument. In this case pynac just passes, a NULL pointer to py_float(), which
lead to crashes.

Changing the declaration from object to PyObject* and checking for NULL fixes
the problem.