Trac #11818: Trac #11818: sage <script> does not run sage-cleaner
When you are running a long computation, you are
probably putting your sage program in a script.sage
file and then run sage script.sage in a screened
terminal. In that case, sage-cleaner is not run.
So if you use @parallel, each fork creates a new
subdirectory $DOT_SAGE/temp/hostname/pid that is
not being cleaned up.