Trac #11987: #11987: Adapt monkeypatch of maxima's "principal" function to follow original behaviour more closely
We patch maxima to throw an error rather than to print "Principal Value" when a divergent
integral is encountered that can be evaluated as a Cauchy Principal Value integral. This
patch brings the behaviour closer to the original in that cases where Maxima silently uses
the principal value, we now do that too.