Trac #12229: Add documentation for contributing to sagenb.
This uses autogenerated documentation from the gitwash project, with slight
modifications on the initial page.

Gitwash: https://github.com/matthew-brett/gitwash

Command to generate documentation:

python gitwash_dumper.py                \
    --repo-name=sagenb                  \
    --github-user=sagemath              \
    --project-url=http://sagemath.org   \
    --project-ml-url=http://groups.google.com/group/sage-notebook   \
    sagenb/                             \
    "Sage Notebook"

Be careful to copy the modifications on the sagenb/index.rst page that tell how
to update the Mercurial and switch to Git.